Hide `set_hd_mode.log` inside $HOME
As a minor annoyance described in #21 the best way to handle it is to make it a hidden file as `xsession-errors.log` and the like. We are not sure if any XDG directory structure is present on first startup and should not rely on it. By hiding it, we make sure it works and does not pollute user interaction with the filesystem.